permission et securisation des commande

+++ main/controller/ControllerCommande.php
@@ -2,14 +2,22 @@
 
 require_once File::build_path(["model", "ModelCommande.php"]);
 require_once File::build_path(["model", "ModelLivre.php"]);
+
 class ControllerCommande {
 
     public static function readAll() {
-        $view = 'list';
-        $pagetitle = 'Liste des Commandes';
-        $controller = 'commande';
-        $tab_co = ModelCommande::selectAll();     //appel au modèle pour gerer la BD
-        require File::build_path(array("view", "view.php"));  //"redirige" vers la vue
+        if (Session::is_admin()) {
+            $view = 'list';
+            $pagetitle = 'Liste des Commandes';
+            $controller = 'commande';
+            $tab_co = ModelCommande::selectAll();     //appel au modèle pour gerer la BD
+            require File::build_path(array("view", "view.php"));  //"redirige" vers la vue
+        } else {
+            echo 'accès restreint';
+            $controller = ('commande');
+            $view = 'error';
+            require (File::build_path(array("view", "view.php")));
+        }
     }
 
     public static function read() {
@@ -17,12 +25,17 @@ class ControllerCommande {
         $pagetitle = 'Commande Details';
         $numco = $_GET['numCommande'];
         $co = ModelCommande::select($numco);
-        $livr = ModelLivre::select($co->getnumLivre());
-        if ($numco == null) {
-            $view = 'error';
-            require File::build_path(array("view", "view.php"));
+        if (Session::is_admin() || Session::is_user($co->getLogin())) {
+            $livr = ModelLivre::select($co->getnumLivre());
+            if ($numco == null) {
+                $view = 'error';
+                require File::build_path(array("view", "view.php"));
+            } else {
+                $view = 'detail';
+                require File::build_path(array("view", "view.php"));
+            }
         } else {
-            $view = 'detail';
+            $view = 'error';
             require File::build_path(array("view", "view.php"));
         }
     }
@@ -50,8 +63,8 @@ class ControllerCommande {
         $pagetitle = 'Liste des commandes';
         require (File::build_path(array("view", "view.php")));
     }
-    
-    public static function validate(){
+
+    public static function validate() {
         if (isset($_SESSION['login'])) {
             if (isset($_COOKIE["panier"]) && !empty(unserialize($_COOKIE["panier"]))) {
                 $panier = unserialize($_COOKIE["panier"]);
@@ -64,23 +77,23 @@ class ControllerCommande {
                     ModelCommande::save($data);
                 }
                 $panier = array();
-                setcookie ("panier", serialize($panier), time()+3600);
+                setcookie("panier", serialize($panier), time() + 3600);
                 $_SESSION["prixPanier"] = 0;
                 $controller = 'commande';
                 $view = 'remerciement';
-                $pagetitle='Merci beaucoup';
-                require File::build_path(array('view','view.php'));
+                $pagetitle = 'Merci beaucoup';
+                require File::build_path(array('view', 'view.php'));
             } else {
                 $controller = 'commande';
                 $view = 'panier';
-                $pagetitle='Panier';
-                require File::build_path(array('view','view.php'));
+                $pagetitle = 'Panier';
+                require File::build_path(array('view', 'view.php'));
             }
-        }else{
+        } else {
             $controller = 'utilisateur';
             $view = 'connect';
-            $pagetitle='Connectez-vous';
-            require File::build_path(array('view','view.php'));
+            $pagetitle = 'Connectez-vous';
+            require File::build_path(array('view', 'view.php'));
         }
     }
 
@@ -94,16 +107,22 @@ class ControllerCommande {
             $controller = ('commande');
             $view = 'error';
             require (File::build_path(array("view", "view.php")));
-        } else {
+        } else if (Session::is_admin() || Session::is_user($co->getLogin())) {
             ModelCommande::delete($numco);
             $controller = ('commande');
             $view = 'deleted';
             $pagetitle = 'Suppression de la commande';
             require (File::build_path(array("view", "view.php")));
+        } else {
+            $pagetitle = 'Commande innexistante';
+            $controller = ('commande');
+            $view = 'error';
+            require (File::build_path(array("view", "view.php")));
         }
     }
 
     public static function update() {
+
         $act = "updated";
         $form = "readonly";
         $pagetitle = 'Mise à jour infos commande';
@@ -120,9 +139,14 @@ class ControllerCommande {
             $controller = ('commande');
             $view = 'error';
             require (File::build_path(array("view", "view.php")));
-        } else {
+        } else if (Session::is_admin()) {
             $controller = 'commande';
             $view = 'update';
+            require (File::build_path(array("view", "view.php")));
+        } else {
+            $pagetitle = 'Commande innexistante';
+            $controller = ('commande');
+            $view = 'error';
             require (File::build_path(array("view", "view.php")));
         }
     }
@@ -137,11 +161,19 @@ class ControllerCommande {
             "numLivre" => $_GET["numLivre"],
         );
         $co = ModelCommande::select($numco);
-        $co->update($data);
-        $controller = "commande";
-        $view = 'updated';
-        require (File::build_path(array("view", "view.php")));
+        if (Session::is_admin()) {
+            $co->update($data);
+            $controller = "commande";
+            $view = 'updated';
+            require (File::build_path(array("view", "view.php")));
+        } else {
+            $pagetitle = 'Commande innexistante';
+            $controller = ('commande');
+            $view = 'error';
+            require (File::build_path(array("view", "view.php")));
+        }
     }
+
     public static function readHistorique() {
         $view = 'list';
         $pagetitle = 'Historique des Commandes';
@@ -149,7 +181,6 @@ class ControllerCommande {
         $tab_co = ModelCommande::selectByLogin();     //appel au modèle pour gerer la BD
         require File::build_path(array("view", "view.php"));  //"redirige" vers la vue
     }
-    
 
 }
